requirements.txt
# Runtime dependencies (top-level).
# For fully reproducible installs (exact versions), use requirements-lock.txt.

stable_baselines3==2.4.1

gym-retro==0.8.0
gymnasium==1.0.0

# Core numerical / logging / plotting
numpy==1.24.4
torch==2.4.1
tensorboard==2.14.0

pandas==2.0.3
matplotlib==3.7.5
opencv-python==4.12.0.88
